各コードのサンプル作成、メインから指定した関数を呼び出して並列化する事ができるようにした。
