+++ code/main_sample.py
@@ -1,0 +1,3 @@
+from module import pool_parallel as po
+from module import random_sample as ras
+po.pool_pal(ras.coin, 4, 10)
